+++ packages/_debian-common/changelog
@@ -4,7 +4,7 @@ libsuperiormysqlpp (0.4.1) UNRELEASED; urgency=medium
   * refactor: Logging system (DRY)
   * feat: Default logger logs errors
 
-  [ MrWarlockX ]
+  [ Michal Merc ]
   * feat: Implement psQuery function
 
  -- Peter Opatril <petr.opatril@firma.seznam.cz>  Fri, 24 May 2019 14:07:13 +0000
